Adds custom 404 page.
Implements a custom 404 page to improve user experience when a page is not found.

It displays a user-friendly message with a link back to the home page and a contact email for error reporting.

The home URL is dynamically set based on the environment, defaulting to the production URL or root during build time, and updated with the window origin on the client side.